remove macro event broker ports feature
move message wrapper tests into models
