+++ crates/documents/Cargo.toml
@@ -63,12 +63,7 @@ outbound = [
   "ports",
 ]
 ports = ["dep:tokio"]
-service = [
-  "dep:entity_access_management",
-  "dep:foreign_entity",
-  "macro_event_broker/ports",
-  "ports",
-]
+service = ["dep:entity_access_management", "dep:foreign_entity", "ports"]
 
 [dependencies]
 # Core
