+++ crates/email/Cargo.toml
@@ -37,7 +37,7 @@ gmail_token = ["dep:authentication_service_client", "dep:redis", "ports"]
 http_client = ["dep:reqwest", "ports"]
 inbound = ["ports"]
 outbound = ["dep:sqlx", "dep:email_db_client", "ports"]
-ports = ["crm/ports", "dep:tokio", "frecency/ports", "macro_event_broker/ports"]
+ports = ["crm/ports", "dep:tokio", "frecency/ports"]
 
 [dependencies]
 ai_toolset = { path = "../ai_toolset", optional = true }
@@ -92,9 +92,6 @@ workspace-hack = { version = "0.1", path = "../workspace-hack" }
 [dev-dependencies]
 crm = { path = "../crm", features = ["outbound"] }
 frecency = { path = "../frecency", features = ["outbound", "postgres"] }
-macro_event_broker = { path = "../macro_event_broker", default-features = false, features = [
-  "ports",
-] }
 macro_db_migrator = { path = "../macro_db_migrator" }
 serde_json = { workspace = true }
 tracing-subscriber = { workspace = true, features = ["env-filter"] }
