+++ crates/macro_event_broker/src/domain/ports.rs
@@ -1,13 +1,8 @@
 //! Port traits defining the broker's inbound and outbound boundaries.
 
-use std::{future::Future, marker::PhantomData};
-
 use macro_event_topics::Topic;
 
-use crate::domain::models::{EventBrokerError, MacroEvent};
-
-#[cfg(test)]
-mod test;
+use crate::domain::models::{EventBrokerError, MacroEvent, MessageWrapper};
 
 /// Inbound port: the public API for sending events through the broker.
 ///
@@ -40,81 +35,10 @@ pub trait EventPublisher: Send + Sync + 'static {
     ) -> impl Future<Output = Result<(), EventBrokerError>> + Send;
 }
 
-/// A broker message paired with the type of event collection it can decode.
-pub struct MessageWrapper<T, M> {
-    inner: T,
-    marker: PhantomData<M>,
-}
-
-impl<T, M> MessageWrapper<T, M> {
-    /// Wraps the transport-specific message.
-    pub fn new(inner: T) -> Self {
-        Self {
-            inner,
-            marker: PhantomData,
-        }
-    }
-
-    /// Borrows the transport-specific message.
-    pub fn inner(&self) -> &T {
-        &self.inner
-    }
-
-    /// Returns the transport-specific message.
-    pub fn into_inner(self) -> T {
-        self.inner
-    }
-}
-
 /// A collection of [`MacroEvent`] types that can decode a broker message.
 pub trait MacroEventCollection: Sized {
     /// Decodes an event from the supplied broker message.
     fn decode<T: MessageParts>(message: &T) -> Result<Self, EventBrokerError>;
-}
-
-/// Declares the [`MacroEvent`] types accepted by a consumer.
-///
-/// The macro creates a `DeclaredMacroEvent` enum with one variant per supplied
-/// event type and implements topic-based decoding for it. Event types must be
-/// identifiers and must implement [`MacroEvent`].
-#[macro_export]
-macro_rules! declare_topics {
-    ($($event:ident),+ $(,)?) => {
-        /// An event decoded from one of the topics declared by `declare_topics!`.
-        pub enum DeclaredMacroEvent {
-            $(
-                #[doc = concat!("A decoded `", stringify!($event), "` event.")]
-                $event($event),
-            )+
-        }
-
-        impl $crate::MacroEventCollection for DeclaredMacroEvent {
-            fn decode<T: $crate::MessageParts>(
-                message: &T,
-            ) -> Result<Self, $crate::EventBrokerError> {
-                $(
-                    if message.topic() == $crate::Topic::as_str(
-                        &<<<$event as $crate::MacroEvent>::EventPayload as $crate::TopicEvent>::Topic as Default>::default(),
-                    ) {
-                        return <$event as $crate::MacroEvent>::decode(
-                            message.key(),
-                            message.payload(),
-                        )
-                        .map(DeclaredMacroEvent::$event);
-                    }
-                )+
-
-                Err($crate::EventBrokerError::UnknownTopic(message.topic().to_owned()))
-            }
-        }
-    };
-}
-
-impl<T: MessageParts, M: MacroEventCollection> MessageWrapper<T, M> {
-    /// Decodes the message into one of the declared event types.
-    pub fn decode_payload(&self) -> Result<M, EventBrokerError> {
-        M::decode(&self.inner)
-    }
 }
 
 /// Broker message fields needed to decode a [`MacroEvent`].
