+++ crates/teams/Cargo.toml
@@ -38,7 +38,6 @@ ports = [
   "dep:model_notifications",
   "dep:notification",
   "dep:tokio",
-  "macro_event_broker/ports",
 ]
 
 [dependencies]
